requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
celery==5.3.4
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
frozenlist==1.4.0
gTTS==2.3.2
magic-filter==1.0.11
multidict==6.0.4
Pillow==10.0.0
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
redis==5.0.0
rembg==2.0.50
SpeechRecognition==3.10.0
urllib3==2.0.4
yarl==1.9.2

openai==0.28.1
fastapi==0.101.1